requirements.txt
pymupdf
pyside6
python3-docx
python3-dotenv
